requirements.txt
Django==4.2.2
django-modeltranslation==0.18.11
django-environ==0.10.0
django-timezone-field==5.0
tzdata==2023.3
django-countries==7.5.1
django-mptt==0.14.0
django-filter==23.2
uuid==1.30
six==1.16.0
celery==5.3.3
redis==5.0.0
channels[daphne]
channels-redis==4.1.0
django-redis==5.3.0
django-debug-toolbar==4.1.0
